woohoo Paper.spec.tsx works
