+++ package.json
@@ -86,6 +86,7 @@
     "@rosskevin/react-docgen": "^3.0.0-beta9",
     "@types/chai": "^4.0.10",
     "@types/classnames": "^2.2.3",
+    "@types/deepmerge": "^1.3.3",
     "@types/enzyme": "^3.1.6",
     "@types/mocha": "^2.2.44",
     "@types/prop-types": "^15.5.2",
