+++ src/Paper/Paper.js
@@ -1,40 +1,10 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import * as classNames from 'classnames';
-import warning = require('warning');
 import withStyles from '../styles/withStyles';
-import { StandardProps } from '..';
+import warning from '../utils/ts-import-workaround/warning';
 
-export type PaperClassKey =
-  | 'root'
-  | 'rounded'
-  | 'shadow0'
-  | 'shadow1'
-  | 'shadow2'
-  | 'shadow3'
-  | 'shadow4'
-  | 'shadow5'
-  | 'shadow6'
-  | 'shadow7'
-  | 'shadow8'
-  | 'shadow9'
-  | 'shadow10'
-  | 'shadow11'
-  | 'shadow12'
-  | 'shadow13'
-  | 'shadow14'
-  | 'shadow15'
-  | 'shadow16'
-  | 'shadow17'
-  | 'shadow18'
-  | 'shadow19'
-  | 'shadow20'
-  | 'shadow21'
-  | 'shadow22'
-  | 'shadow23'
-  | 'shadow24'
-  ;
-export const styles = withStyles<PaperClassKey>((theme) => {
+export const styles = theme => {
   const shadows = {};
   theme.shadows.forEach((shadow, index) => {
     shadows[`shadow${index}`] = {
@@ -51,18 +21,9 @@ export const styles = withStyles<PaperClassKey>((theme) => {
     },
     ...shadows,
   };
-},  { name: 'MuiPaper' })
+};
 
-export interface PaperProps extends StandardProps<
-  React.HTMLAttributes<HTMLDivElement>,
-  PaperClassKey
-  > {
-  component?: string | React.ComponentType<PaperProps>;
-  elevation?: number;
-  square?: boolean;
-}
-
-const Paper: React.StatelessComponent = (props: PaperProps) => {
+function Paper(props) {
   const {
     classes,
     className: classNameProp,
@@ -124,4 +85,4 @@ Paper.defaultProps = {
   square: false,
 };
 
-export default (Paper);
+export default withStyles(styles, { name: 'MuiPaper' })(Paper);
