+++ src/utils/ts-import-workaround/deepmerge.ts
@@ -1,0 +1,3 @@
+import deepmerge = require('deepmerge'); // < 1kb payload overhead when lodash/merge is > 3kb.
+
+export default deepmerge;
