+++ src/utils/ts-import-workaround/warning.ts
@@ -1,0 +1,3 @@
+import warning = require('warning');
+
+export default warning
